requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
click==7.1.2
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.12.1
djongo==1.3.3
factory-boy==3.0.1
Faker==4.1.3
Markdown==3.2.2
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.2
pathspec==0.8.0
psycopg2==2.8.6
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.9.27
six==1.15.0
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
